drm: mwv206: Support JMGPU JM7200
Add Jemoic JM7200 GPU driver support.

mwv206 kernel driver source for Jiajing project from commit f904753
("Import 1.1.2.20200929 from Jingjia") in mwv206-kernel repo from
FlyGoat. While he may not be the original author of the driver (the
copyright for the original driver most likely belongs to Jemoic
Corporation), we are nonetheless immensely grateful for his efforts.
His contribution provided us with a channel to obtain the driver's
source code and port it into the kernel.

This was a very basic and rough driver porting effort. The original
driver's code quality was extremely poor, making it difficult to
even compile. The sole purpose of this port was to enable the driver
to compile correctly into a .ko file within the kernel source tree
for testing purposes.

Other parts of the code have been kept as close to their original
state as possible.

We hope to refine it further if an opportunity arises in the future.

During the porting process, some incompatible changes were introduced,
such as masking out atrocious build-process macros like _MWV206_CPU_
and _MWV206_OS_. The advantage of this approach was to expedite the
porting work and allow for verification on x86 platforms. The drawback
is that it might cause issues with the driver on ARM and MIPS
architectures.

Additionally, it appears this driver doesn't support architectures
other than x86, ARM64, and MIPS64EL, which is incorrect. Hopefully,
if we're lucky, someone will eventually step forward to refactor
this mess of code.

Origin: https://github.com/FlyGoat/mwv206-kernel
Author: rfshen <jjwgpu@jingjiamicro.com>
Cc: Jiaxun Yang <jiaxun.yang@flygoat.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>